Powershell framework to find and remove files by name on Windows Active Directory domain.
The file variables can be modified to suit your needs, same for the paths of where to find the files, along with the operating system where statements.
Please test this script on a test environment so that you know what to expect. This does work successfully and only removes the files you tell it to.
This script was purpose built to cleanup old utilities on servers built with a template.

Plan to add hash search capability and reporting in future update.